raphael: lights: Change LED behavior
Raphael has 2 LED lights. One on left of the popup (Name is green) and on on the right of popup (Name is blue). Don't know why it is named so weirdly. According to stock behavior, only the left LED should pulse on notification or battery as indicator since left side has a small circular cutout to pass the light. To mimick that, I removed the the right LED defination, since it is never used for notifications or battery light in stock, and if is enabled, causes a light bleed from other side of popup.

naveenjohnsonv: adapt for new tree and removed LEDs not present on our device
Signed-off-by: naveenjohnsonv <14140949+naveenjohnsonv@users.noreply.github.com>